+++ excellent/functions/functions.go
@@ -1468,6 +1468,7 @@ func ParseJSON(env utils.Environment, text types.XText) types.XValue {
 //
 //   @(json("string")) -> "string"
 //   @(json(10)) -> 10
+//   @(json(null)) -> null
 //   @(json(contact.uuid)) -> "5d76d86b-3bb9-4d5a-b822-c9d86f5d8e4f"
 //
 // @function json(value)
@@ -1801,6 +1802,7 @@ func Length(env utils.Environment, value types.XValue) types.XValue {
 //   @(default(array(1, 2), "value")) -> [1, 2]
 //   @(default(array(), "value")) -> value
 //   @(default(datetime("invalid-date"), "today")) -> today
+//   @(default(format_urn("invalid-urn"), "ok")) -> ok
 //
 // @function default(value, default)
 func Default(env utils.Environment, value types.XValue, def types.XValue) types.XValue {