+++ excellent/functions/functions_test.go
@@ -277,6 +277,7 @@ func TestFunctions(t *testing.T) {
 		{"format_number", dmy, []types.XValue{ERROR}, ERROR},
 		{"format_number", dmy, []types.XValue{}, ERROR},
 
+		{"format_urn", dmy, []types.XValue{xs("tel:+14132378053")}, xs("(413) 237-8053")},
 		{"format_urn", dmy, []types.XValue{xs("tel:+250781234567")}, xs("0781 234 567")},
 		{"format_urn", dmy, []types.XValue{xs("twitter:134252511151#billy_bob")}, xs("billy_bob")},
 		{"format_urn", dmy, []types.XValue{xs("NOT URN")}, ERROR},
@@ -303,6 +304,7 @@ func TestFunctions(t *testing.T) {
 		{"join", dmy, []types.XValue{types.NewXArray(xs("1"))}, ERROR},
 
 		{"json", dmy, []types.XValue{xs("hello")}, xs(`"hello"`)},
+		{"json", dmy, []types.XValue{nil}, xs(`null`)},
 		{"json", dmy, []types.XValue{ERROR}, ERROR},
 
 		{"left", dmy, []types.XValue{xs("hello"), xs("2")}, xs("he")},